Convert from Float64 to Float32

+++ src/structs/ApertureApodization.jl
@@ -5,16 +5,16 @@ export ApertureApodization
     focus::AbstractScan = Scan()
 
     window::Window.WindowType = Window.None
-    f_number::Vector{Float64} = [1.0, 1.0]
+    f_number::Vector{Float32} = [1.0, 1.0]
 
     MLA::Integer = 1
     MLA_overlap::Integer = 0
     
-    tilt::Vector{Float64} = [0.0, 0.0]
-    minimum_aperature::Vector{Float64} = [1e-3, 1e-3]
-    maximum_aperature::Vector{Float64} = [10, 10]
+    tilt::Vector{Float32} = [0.0, 0.0]
+    minimum_aperature::Vector{Float32} = [1e-3, 1e-3]
+    maximum_aperature::Vector{Float32} = [10, 10]
 
-    apodization_vector::Vector{Float64} = []
+    apodization_vector::Vector{Float32} = []
     origin::Point = Point()
 end
