+++ src/structs/CurvilinearMatrixArray.jl
@@ -2,7 +2,7 @@ export CurvilinearMatrixArray
 
 @kwdef mutable struct CurvilinearMatrixArray <: AbstractProbe
     origin::Point = Point()
-    geometry::Array{Float64, 2} = Array{Float64, 2}(undef, 0, 7)
+    geometry::Array{Float32, 2} = Array{Float32, 2}(undef, 0, 7)
 
-    radius_x::Float64 = 0.0
+    radius_x::Float32 = 0.0
 end
